ngUnitTestingDemo/src/app/hero-detail/hero-detail.component.ts
import { Component, OnInit, Input } from "@angular/core";
import { ActivatedRoute } from "@angular/router";
import { Location } from "@angular/common";

import { Hero } from "../hero";
import { HeroService } from "../hero.service";

@Component({
  selector: "app-hero-detail",
  templateUrl: "./hero-detail.component.html",
  styleUrls: ["./hero-detail.component.css"],
})
export class HeroDetailComponent implements OnInit {
  @Input() hero: Hero;

  constructor(
    private route: ActivatedRoute,
    private heroService: HeroService,
    private location: Location
  ) {}

  ngOnInit(): void {
    console.log("Inside ngOnInit of hero detail component");
    this.getHero();
  }

  getHero(): void {
    const id = +this.route.snapshot.paramMap.get("id");
    this.heroService.getHero(id).subscribe((hero) => (this.hero = hero));
  }

  goBack(): void {
    console.log("back called");
    this.location.back();
  }

  save(): void {
    console.log("Save called");
    this.heroService.updateHero(this.hero).subscribe(() => {
      console.log("Saved hero");
      this.goBack();
    });
  }
}
